Filter a list of predictions cards based on attributes

Filter a list of predictions cards based on attributes

filter_predictions(
  predictions_cards,
  name_of_forecaster = NULL,
  response_data_source = NULL,
  response_signal = NULL,
  forecast_date = NULL,
  incidence_period = NULL,
  ahead = NULL,
  geo_type = NULL
)

Arguments

predictions_cards
List of predictions cards, as created by get_predictions().

name_of_forecaster
String indicating name of the forecaster.

response_data_source
String indicating response data source.

response_signal
String indicating response signal.

forecast_date
Date object or string of the form "YYYY-MM-DD", indicating date on which forecasts will be made about some period (e.g., epiweek). For example, if forecast_date = "2020-05-11", incidence_period = "day", and ahead = 3, then, we'd be making forecasts for "2020-05-14".

incidence_period
String indicating the incidence period, either "epiweek" or "day".

ahead
Vector of (one or more) integers. How many epiweeks/days ahead are you forecasting? If incidence_period = "epiweek" and forecast_date is Sunday or Monday, then ahead = 1 means the epiweek that includes the forecast date; if forecast_date falls on a Tuesday through Saturday, then it is the following epiweek. If incidence_period = "day", then ahead = 1 means the day after forecast date.

geo_type
String indicating geographical type, one of "county", "dma", "hrr", "msa", or "state".