Get predictions from a forecaster on the COVID Hub

Simply converts the predictions of forecasters submitting to the COVID Hub to the format of a predictions card, so it can be easily evaluated and compared.

get_covidhub_predictions(covid_hub_forecaster_name, forecast_dates = NULL, ...)

Arguments

covid_hub_forecaster_name
String indicating of the forecaster (matching what it is called on the COVID Hub).

forecast_dates
Vector of Date objects (or strings of the form "YYYY-MM-DD") indicating dates on which forecasts will be made. If NULL, the default, then all currently available forecast dates from the given forecaster in the COVID Hub will be used.

...
Additional parameters to be passed to filter_predictions().

Value

List of predictions cards.

Details

For now, this function only supports (i) incident not cumulative predictions and (ii) epiweek not daily incidence period predictions.

See also