Get predictions cards for a single date

Get predictions cards for a single date

get_predictions_single_date(
  forecaster,
  name_of_forecaster,
  signals,
  forecast_date,
  incidence_period,
  ahead,
  geo_type,
  geo_values,
  ...
)

Arguments

forecaster
Function that outputs a tibble with columns ahead, location, probs, quantiles. The quantiles column gives the predictive quantiles of the forecast distribution for that location and ahead.

name_of_forecaster
String indicating name of the forecaster.

signals
Tibble with columns data_source and signal that specifies which variables from the COVIDcast API will be used by forecaster. Each row of signals represents a separate signal, and first row is taken to be the response. If using incidence_period = "epiweek", the response should be something for which summing daily values over an epiweek makes sense (e.g., counts or proportions but not log(counts) or log(proportions)). Available data sources and signals are documented in the COVIDcast signal documentation. An optional column start_day can also be included. This can be a Date object or string in the form "YYYY-MM-DD", indicating the earliest date of data needed from that data source. Importantly, start_day can also be a function (represented as a list column) that takes a forecast date and returns a start date for model training (again, Date object or string in the form "YYYY-MM-DD"). The latter is useful when the start date should be computed dynamically from the forecast date (e.g., when forecaster only trains on the most recent 4 weeks of data).

forecast_date
Date object or string of the form "YYYY-MM-DD", indicating date on which forecasts will be made about some period (e.g., epiweek). For example, if forecast_date = "2020-05-11", incidence_period = "day", and ahead = 3, then, we'd be making forecasts for "2020-05-14".

incidence_period
String indicating the incidence period, either "epiweek" or "day".

ahead
Vector of (one or more) integers. How many epiweeks/days ahead are you forecasting? If incidence_period = "epiweek" and forecast_date is Sunday or Monday, then ahead = 1 means the epiweek that includes the forecast date; if forecast_date falls on a Tuesday through Saturday, then it is the following epiweek. If incidence_period = "day", then ahead = 1 means the day after forecast date.

geo_type
String indicating geographical type, one of "county", "dma", "hrr", "msa", or "state".

geo_values
Vector or list of locations whose data will be passed to forecaster. The default, "*", fetches all that is available in the API.

...
Additional arguments to be passed to forecaster().