Plot calibration curves

Plot calibration curves

plot_calibration(
  predictions_cards,
  type = c("wedgeplot", "traditional"),
  facet_rows = NULL,
  facet_cols = NULL,
  grp_vars = c("forecaster", "forecast_date", "ahead"),
  avg_vars = c("geo_value"),
  geo_type = c("county", "hrr", "msa", "dma", "state", "hhs", "nation"),
  backfill_buffer = 0
)

Arguments

predictions_cards
Either one predictions_card or several joined together using bind_rows().

type
One of "wedgeplot" or "traditional"

facet_rows
A variable name to facet data over. Creates a separate row of plots for each value of specified variable. Can be used with facet_cols to create a grid of plots. Should be passed to plot_calibration when customized.

facet_cols
Same as facet_rows, but with columns.

grp_vars
variables over which to compare calibration. These determines the color of the lines and faceting depending on type

avg_vars
variables over which we average to determine the calibration.

geo_type
String indicating geographical type, such as "county", or "state". See the COVIDcast Geographic Coding documentation for available options.

backfill_buffer
How many days until response is deemed trustworthy enough to be taken as correct?.