Get predictions

For each of the provided forecast dates, runs a forecaster using the data that would have been available as of that given forecast date. Returns a list of "predictions cards", where each list element corresponds to a different forecast date. A predictions card is a data frame giving the forecast distributions of a given forecaster for a given forecast task. A forecast task is specified by the forecast date, ahead, response, incidence period, and geo type (e.g., 1-epiweek-ahead death forecasting at the state level with predictions made using the information as of September 14).

get_predictions(
  forecaster,
  name_of_forecaster,
  signals,
  forecast_dates,
  incidence_period,
  ahead,
  geo_type,
  geo_values = "*",
  apply_corrections = NULL,
  signal_aggregation = c("list", "wide", "long"),
  signal_aggregation_dt = NULL,
  ...
)

Arguments

forecaster
Function that outputs a tibble with columns ahead, geo_value, quantile, and value. The quantile column gives the probabilities associated with quantile forecasts for that location and ahead. If your forecaster produces point forecasts, then set quantile=NA.

name_of_forecaster
String indicating name of the forecaster.

signals
Tibble with columns data_source and signal that specifies which variables from the COVIDcast API will be used by forecaster. Each row of signals represents a separate signal, and first row is taken to be the response. If using incidence_period = "epiweek", the response should be something for which summing daily values over an epiweek makes sense (e.g., counts or proportions but not log(counts) or log(proportions)). Available data sources and signals are documented in the COVIDcast signal documentation. An optional column start_day can also be included. This can be a Date object or string in the form "YYYY-MM-DD", indicating the earliest date of data needed from that data source. Importantly, start_day can also be a function (represented as a list column) that takes a forecast date and returns a start date for model training (again, Date object or string in the form "YYYY-MM-DD"). The latter is useful when the start date should be computed dynamically from the forecast date (e.g., when forecaster only trains on the most recent 4 weeks of data).

forecast_dates
Vector of Date objects (or strings of the form "YYYY-MM-DD") indicating dates on which forecasts will be made.

incidence_period
String indicating the incidence period, either "epiweek" or "day".

ahead
Vector of (one or more) integers. How many epiweeks/days ahead are you forecasting? If incidence_period = "epiweek" and forecast_date is Sunday or Monday, then ahead = 1 means the epiweek that includes the forecast date; if forecast_date falls on a Tuesday through Saturday, then it is the following epiweek. If incidence_period = "day", then ahead = 1 means the day after forecast date.

geo_type
String indicating geographical type, one of "county", "dma", "hrr", "msa", or "state".

geo_values
Vector or list of locations whose data will be passed to forecaster. The default, "*", fetches all that is available in the API.

apply_corrections
an optional function that applies data corrections to the signals. Input is a data frame or list as returned as df <- covidcast::download_signals(). The returned object should be of the type expected by your forecaster. This function will be called as apply_corrections(df, ...) using any named arguments passed in ... to get_predictions().

signal_aggregation
this and the next argument control the type of data your forecaster expects to receive from covidcast. By default, different signals are passed in "list" format. But you may alternatively request "wide" or "long". See covidcast::covidcast_signals() and covidcast::aggregate_signals() for more details.

signal_aggregation_dt
for any data format, covidcast::aggregate_signals() can perform leading and lagging for you. See that documentation for more details.

...
Additional named arguments to be passed to forecaster() and apply_corrections()

Value

Long data frame of forecasts with a class of predictions_cards. The first 4 columns are the same as those returned by the forecaster. The remainder specify the prediction task, 10 columns in total: ahead, geo_value, quantile, value, forecaster, forecast_date, data_source, signal, target_end_date, and incidence_period. Here data_source and signal correspond to the response variable only.

Examples

baby_predictions = get_predictions(
  baseline_forecaster, "baby",
  tibble::tibble(
    data_source=c("jhu-csse", "usa-facts"),
    signal = c("deaths_incidence_num","confirmed_incidence_num"),
    start_day="2020-08-15"), 
  "2020-10-01","epiweek", 1:4, 
  "state", "mi", signal_aggregation="long")

#> Downloading deaths_incidence_num from covidcast for period from 2020-08-15 to 2020-10-01.

#> Downloading confirmed_incidence_num from covidcast for period from 2020-08-15 to 2020-10-01.

#> Warning: Data not fetched for the following days: 2020-10-01

#> Warning: Data not fetched for the following days: 2020-10-01